🖥️fix(controller): Fix condition for /novels DELETE endpoint
- Return 404 when the list of ids is empty.
- Return 404 when at least one id in the list does not exist in the db.

+++ glitch-novels-backend/resources-server/src/main/java/com/namdp/glitch_novels/resources_server/controllers/NovelController.java
@@ -40,17 +40,23 @@ public class NovelController {
 
   @DeleteMapping("/novels")
   public ResponseEntity<Map<String, Object>> deleteNovels(@RequestBody List<Integer> ids) {
-    List<NovelDTO> deletedNovels = novelService.deleteNovelsByIds(ids);
-
     Map<String, Object> responseBody = new HashMap<>();
-    if (deletedNovels.size() == 0) {
+
+    if (ids.isEmpty()) {
+      responseBody.put("message", "The list of ids cannot be empty.");
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body(responseBody);
+    }
+
+    // Only delete if all the novels with the given ids exist in db.
+    if (!novelService.checkIfAllNovelsExist(ids)) {
       responseBody.put("message", "Some or all of the provided novel ids could not be found.");
       return ResponseEntity.status(HttpStatus.NOT_FOUND).body(responseBody);
-    } else {
-      responseBody.put("message", "The novels have been successfully deleted.");
-      responseBody.put("deletedNovel", deletedNovels);
-      return ResponseEntity.ok(responseBody);
     }
+
+    List<NovelDTO> deletedNovels = novelService.deleteNovelsByIds(ids);
+    responseBody.put("message", "The novels have been successfully deleted.");
+    responseBody.put("deletedNovel", deletedNovels);
+    return ResponseEntity.ok(responseBody);
   }
 
   @GetMapping("/novel/{id}")
